libvirt: fix detach volume race condition
When a volume is detached, the according share file is
detached if no longer used. This can cause issues if at the same
time, a different volume stored on the same share is being
attached as the according disk image will not be available.
This patch fixes the issue by tracking share usage.

Change-Id: I7b7ad794b1e15dcd63344e01a520d6c595a73d70
Closes-Bug: #1421550
Closes-Bug: #1461081