encryptors: Use a mangled passphrase if permission is denied
Prior to Ib563b0ea the passphrase used by CryptsetupEncryptor and
LuksEncryptor had any leading zeros per hexadecimal digit removed, for
example 0x04 or 04 would turn into 0x4 or 4. As a result any volume
encrypted prior to the release of Newton used a modified passphrase that
was different to that stored by the key manager being used in the
environment.

In order to allow these volumes to still be used permission denied
errors are now caught and a second attempt is made to access the volume
using a mangled passphrase, with all leading zeros removed from each
hexadecimal number.

Change-Id: I7096463c5eba951dd6322ee6965435e877ca0371
Closes-bug: #1633518