libvirt: Reconnect volumes and encryptors during a hard reboot
When hard rebooting an instance the Libvirt driver would previously
leave volumes connected and encryptors attached on the host. This could
lead to failures when recreating the domain as Nova attempts to connect
to volumes that are already connected and have encryptors attached.

To avoid this a full destroy is now called for when hard rebooting an
instance. This detaches any encryptors and disconnects the volumes from
the host allowing the volumes to be reconnected and encryptors
reattached when the instance is redefined.

Closes-bug: #1597234
Change-Id: Id188d48609f3d22d14e16c7f6114291d547a8986