WIP Add regression test for bug #1843639
This change adds a simple functional change to reproduce bug #1843639
where exceptions encountered during post_live_migration on the source
compute node result in the overall failure of a live migration.

ServiceFixture is also updated to support multiple Libvirt based n-cpu
services being launched with unique hostnames. This is required as the
libvirt fixture itself hardcodes a single hostname of 'compute1' as the
return to getHostname calls. This in turn results in each compute
attempting to report as the same node to Placement when starting up.

WIP as there's something going on with the mocking when I attempt to
remove wait_till_active_or_timeout.

Related-Bug: #1843639
Change-Id: I6e0a1b977b2e5eef37309e1494e94568cf187bd8